+++ build_runner/lib/src/generate/build.dart
@@ -45,6 +45,10 @@ import 'watch_impl.dart' as watch_impl;
 /// will be created for each value in the map which contains all original
 /// sources and built sources contained in the provided path.
 ///
+/// If [outputSymlinksOnly] is `true`, then the merged output directories will
+/// contain only symlinks, which is much faster but not generally suitable for
+/// deployment.
+///
 /// If [verbose] is `true` then verbose logging will be enabled. This changes
 /// the default [logLevel] to [Level.ALL] and removes stack frame folding, among
 /// other things.
@@ -61,6 +65,7 @@ Future<BuildResult> build(List<BuilderApplication> builders,
     Stream terminateEventStream,
     bool enableLowResourcesMode,
     Map<String, String> outputMap,
+    bool outputSymlinksOnly,
     bool trackPerformance,
     bool skipBuildScriptCheck,
     bool verbose,
@@ -71,7 +76,12 @@ Future<BuildResult> build(List<BuilderApplication> builders,
   builderConfigOverrides ??= const {};
   packageGraph ??= new PackageGraph.forThisPackage();
   var environment = new OverrideableEnvironment(
-      new IOEnvironment(packageGraph, assumeTty),
+      new IOEnvironment(
+        packageGraph,
+        assumeTty: assumeTty,
+        outputMap: outputMap,
+        outputSymlinksOnly: outputSymlinksOnly,
+      ),
       reader: reader,
       writer: writer,
       onLog: onLog ?? stdIOLogListener(assumeTty: assumeTty, verbose: verbose));
@@ -84,7 +94,6 @@ Future<BuildResult> build(List<BuilderApplication> builders,
     overrideBuildConfig:
         await findBuildConfigOverrides(packageGraph, configKey),
     enableLowResourcesMode: enableLowResourcesMode,
-    outputMap: outputMap,
     trackPerformance: trackPerformance,
     verbose: verbose,
     buildDirs: buildDirs,
@@ -148,6 +157,7 @@ Future<ServeHandler> watch(List<BuilderApplication> builders,
         Stream terminateEventStream,
         bool enableLowResourcesMode,
         Map<String, String> outputMap,
+        bool outputSymlinksOnly,
         bool trackPerformance,
         bool skipBuildScriptCheck,
         bool verbose,
@@ -171,6 +181,7 @@ Future<ServeHandler> watch(List<BuilderApplication> builders,
       terminateEventStream: terminateEventStream,
       enableLowResourcesMode: enableLowResourcesMode,
       outputMap: outputMap,
+      outputSymlinksOnly: outputSymlinksOnly,
       trackPerformance: trackPerformance,
       skipBuildScriptCheck: skipBuildScriptCheck,
       verbose: verbose,