+++ build_runner_core/lib/src/asset/build_cache.dart
@@ -8,16 +8,24 @@ import 'package:glob/glob.dart';
 import '../asset_graph/graph.dart';
 import '../asset_graph/node.dart';
 import '../util/constants.dart';
+import 'reader.dart';
 import 'writer.dart';
 
 /// Wraps an [AssetReader] and translates reads for generated files into reads
 /// from the build cache directory
 class BuildCacheReader implements AssetReader {
-  final AssetReader _delegate;
   final AssetGraph _assetGraph;
+  final AssetReader _delegate;
   final String _rootPackage;
 
-  BuildCacheReader(this._delegate, this._assetGraph, this._rootPackage);
+  BuildCacheReader._(this._delegate, this._assetGraph, this._rootPackage);
+
+  factory BuildCacheReader(
+          AssetReader delegate, AssetGraph assetGraph, String rootPackage) =>
+      delegate is PathProvidingAssetReader
+          ? new _PathProvidingBuildCacheReader._(
+              delegate, assetGraph, rootPackage)
+          : new BuildCacheReader._(delegate, assetGraph, rootPackage);
 
   @override
   Future<bool> canRead(AssetId id) =>
@@ -41,9 +49,24 @@ class BuildCacheReader implements AssetReader {
       'Asset globbing should be done per phase with the AssetGraph');
 }
 
+class _PathProvidingBuildCacheReader extends BuildCacheReader
+    implements PathProvidingAssetReader {
+  @override
+  PathProvidingAssetReader get _delegate =>
+      super._delegate as PathProvidingAssetReader;
+
+  _PathProvidingBuildCacheReader._(PathProvidingAssetReader delegate,
+      AssetGraph assetGraph, String rootPackage)
+      : super._(delegate, assetGraph, rootPackage);
+
+  @override
+  String pathTo(AssetId id) =>
+      _delegate.pathTo(_cacheLocation(id, _assetGraph, _rootPackage));
+}
+
 class BuildCacheWriter implements RunnerAssetWriter {
-  final RunnerAssetWriter _delegate;
   final AssetGraph _assetGraph;
+  final RunnerAssetWriter _delegate;
   final String _rootPackage;
 
   BuildCacheWriter(this._delegate, this._assetGraph, this._rootPackage);
@@ -51,12 +74,14 @@ class BuildCacheWriter implements RunnerAssetWriter {
   @override
   Future writeAsBytes(AssetId id, List<int> content) => _delegate.writeAsBytes(
       _cacheLocation(id, _assetGraph, _rootPackage), content);
+
   @override
   Future writeAsString(AssetId id, String content,
           {Encoding encoding = utf8}) =>
       _delegate.writeAsString(
           _cacheLocation(id, _assetGraph, _rootPackage), content,
           encoding: encoding);
+
   @override
   Future delete(AssetId id) =>
       _delegate.delete(_cacheLocation(id, _assetGraph, _rootPackage));