+++ build_runner_core/lib/src/environment/build_environment.dart
@@ -4,10 +4,13 @@
 
 import 'dart:async';
 
+import 'package:build/build.dart';
 import 'package:logging/logging.dart';
 
 import '../asset/reader.dart';
 import '../asset/writer.dart';
+import '../generate/build_result.dart';
+import '../generate/finalized_assets_view.dart';
 
 /// Utilities to interact with the environment in which a build is running.
 ///
@@ -29,6 +32,20 @@ abstract class BuildEnvironment {
   /// If this environmment is non-interactive (such as when running in a test)
   /// this method should throw [NonInteractiveBuildException].
   Future<int> prompt(String message, List<String> choices);
+
+  /// Invoked after each build, can modify the [BuildResult] in any way, even
+  /// converting it to a failure.
+  ///
+  /// The [finalizedAssetsView] can only be used until the returned [Future]
+  /// completes, it will expire afterwords since it can no longer guarantee a
+  /// consistent state.
+  ///
+  /// By default this returns the original result.
+  ///
+  /// Any operation may be performed, as determined by environment.
+  Future<BuildResult> finalizeBuild(BuildResult buildResult,
+          FinalizedAssetsView finalizedAssetsView, AssetReader assetReader) =>
+      new Future.value(buildResult);
 }
 
 /// Thrown when the build attempts to prompt the users but no prompt is