+++ build_runner_core/lib/src/environment/create_merged_dir.dart
@@ -10,10 +10,9 @@ import 'package:logging/logging.dart';
 import 'package:path/path.dart' as p;
 import 'package:pool/pool.dart';
 
-import '../asset_graph/graph.dart';
-import '../asset_graph/node.dart';
-import '../asset_graph/optional_output_tracker.dart';
+import '../asset/reader.dart';
 import '../environment/build_environment.dart';
+import '../generate/finalized_assets_view.dart';
 import '../logging/logging.dart';
 import '../package_graph/package_graph.dart';
 
@@ -29,14 +28,21 @@ const _manifestSeparator = '\n';
 /// Returns whether it succeeded or not.
 Future<bool> createMergedOutputDirectories(
     Map<String, String> outputMap,
-    AssetGraph assetGraph,
     PackageGraph packageGraph,
-    AssetReader reader,
     BuildEnvironment environment,
-    OptionalOutputTracker optionalOutputTracker) async {
+    AssetReader reader,
+    FinalizedAssetsView finalizedAssetsView,
+    bool outputSymlinksOnly) async {
+  if (outputSymlinksOnly && reader is! PathProvidingAssetReader) {
+    _logger.severe(
+        'The current environment does not support symlinks, but symlinks were '
+        'requested.');
+    return false;
+  }
+
   for (var output in outputMap.keys) {
-    if (!await _createMergedOutputDir(output, outputMap[output], assetGraph,
-        packageGraph, reader, environment, optionalOutputTracker)) {
+    if (!await _createMergedOutputDir(output, outputMap[output], packageGraph,
+        environment, reader, finalizedAssetsView, outputSymlinksOnly)) {
       _logger.severe('Unable to create merged directory for $output.\n'
           'Choose a different directory or delete the contents of that '
           'directory.');
@@ -47,14 +53,13 @@ Future<bool> createMergedOutputDirectories(
 }
 
 Future<bool> _createMergedOutputDir(
-  String outputPath,
-  String root,
-  AssetGraph assetGraph,
-  PackageGraph packageGraph,
-  AssetReader reader,
-  BuildEnvironment environment,
-  OptionalOutputTracker optionalOutputTracker,
-) async {
+    String outputPath,
+    String root,
+    PackageGraph packageGraph,
+    BuildEnvironment environment,
+    AssetReader reader,
+    FinalizedAssetsView finalizedOutputsView,
+    bool symlinkOnly) async {
   var outputDir = new Directory(outputPath);
   var outputDirExists = await outputDir.exists();
   if (outputDirExists) {
@@ -62,8 +67,7 @@ Future<bool> _createMergedOutputDir(
     if (!result) return result;
   }
 
-  var outputAssets = new Set<AssetId>();
-  var originalOutputAssets = new Set<AssetId>();
+  var outputAssets = <AssetId>[];
 
   await logTimedAsync(_logger, 'Creating merged output dir `$outputPath`',
       () async {
@@ -71,15 +75,10 @@ Future<bool> _createMergedOutputDir(
       await outputDir.create(recursive: true);
     }
 
-    await Future.wait(assetGraph.allNodes.map((node) async {
-      if (_shouldSkipNode(node, root, optionalOutputTracker)) {
-        return;
-      }
-      originalOutputAssets.add(node.id);
-      node.lastKnownDigest ??= await reader.digest(node.id);
-      outputAssets.add(
-          await _writeAsset(node.id, outputDir, root, packageGraph, reader));
-    }));
+    outputAssets.addAll(await Future.wait(finalizedOutputsView
+        .allAssets(rootDir: root)
+        .map((id) => _writeAsset(
+            id, outputDir, root, packageGraph, reader, symlinkOnly))));
 
     var packagesFileContent = packageGraph.allPackages.keys
         .map((p) => '$p:packages/$p/')
@@ -89,8 +88,7 @@ Future<bool> _createMergedOutputDir(
     outputAssets.add(packagesAsset);
 
     if (root == null) {
-      for (var dir in _findRootDirs(
-          outputPath, assetGraph, packageGraph, optionalOutputTracker)) {
+      for (var dir in _findRootDirs(outputAssets, outputPath)) {
         var link = new Link(p.join(outputDir.path, dir, 'packages'));
         if (!link.existsSync()) {
           link.createSync(p.join('..', 'packages'), recursive: true);
@@ -109,14 +107,10 @@ Future<bool> _createMergedOutputDir(
   return true;
 }
 
-Set<String> _findRootDirs(String outputPath, AssetGraph assetGraph,
-    PackageGraph packageGraph, OptionalOutputTracker optionalOutputTracker) {
+Set<String> _findRootDirs(Iterable<AssetId> allAssets, String outputPath) {
   var rootDirs = new Set<String>();
-  for (var node in assetGraph.packageNodes(packageGraph.root.name)) {
-    if (_shouldSkipNode(node, null, optionalOutputTracker)) {
-      continue;
-    }
-    var parts = p.url.split(node.id.path);
+  for (var id in allAssets) {
+    var parts = p.url.split(id.path);
     if (parts.length == 1) continue;
     var dir = parts.first;
     if (dir == outputPath || dir == 'lib') continue;
@@ -125,30 +119,8 @@ Set<String> _findRootDirs(String outputPath, AssetGraph assetGraph,
   return rootDirs;
 }
 
-bool _shouldSkipNode(
-    AssetNode node, String root, OptionalOutputTracker optionalOutputTracker) {
-  if (!node.isReadable) return true;
-  if (node.isDeleted) return true;
-  if (root != null &&
-      !(node.id.path.startsWith('lib/')) &&
-      !p.isWithin(root, node.id.path)) {
-    return true;
-  }
-  if (node is InternalAssetNode) return true;
-  if (node is GeneratedAssetNode) {
-    if (!node.wasOutput ||
-        node.isFailure ||
-        node.state != GeneratedNodeState.upToDate) {
-      return true;
-    }
-    return !optionalOutputTracker.isRequired(node.id);
-  }
-  if (node.id.path == '.packages') return true;
-  return false;
-}
-
 Future<AssetId> _writeAsset(AssetId id, Directory outputDir, String root,
-    PackageGraph packageGraph, AssetReader reader) {
+    PackageGraph packageGraph, AssetReader reader, bool symlinkOnly) {
   return _descriptorPool.withResource(() async {
     String assetPath;
     if (id.path.startsWith('lib/')) {
@@ -164,7 +136,15 @@ Future<AssetId> _writeAsset(AssetId id, Directory outputDir, String root,
 
     var outputId = new AssetId(packageGraph.root.name, assetPath);
     try {
-      await _writeAsBytes(outputDir, outputId, await reader.readAsBytes(id));
+      if (symlinkOnly) {
+        await new Link(_filePathFor(outputDir, id)).create(
+            // We assert at the top of `createMergedOutputDirectories` that the
+            // reader implements this type when requesting symlinks.
+            (reader as PathProvidingAssetReader).pathTo(id),
+            recursive: true);
+      } else {
+        await _writeAsBytes(outputDir, outputId, await reader.readAsBytes(id));
+      }
     } on AssetNotFoundException catch (e, __) {
       if (p.basename(id.path).startsWith('.')) {
         _logger.fine('Skipping missing hidden file ${id.path}');
@@ -188,6 +168,10 @@ Future<void> _writeAsString(Directory outputDir, AssetId id, String contents) =>
     _fileFor(outputDir, id).then((file) => file.writeAsString(contents));
 
 Future<File> _fileFor(Directory outputDir, AssetId id) {
+  return new File(_filePathFor(outputDir, id)).create(recursive: true);
+}
+
+String _filePathFor(Directory outputDir, AssetId id) {
   String relativePath;
   if (id.path.startsWith('lib')) {
     relativePath =
@@ -195,7 +179,7 @@ Future<File> _fileFor(Directory outputDir, AssetId id) {
   } else {
     relativePath = id.path;
   }
-  return new File(p.join(outputDir.path, relativePath)).create(recursive: true);
+  return p.join(outputDir.path, relativePath);
 }
 
 /// Checks for a manifest file in [outputDir] and deletes all referenced files.