+++ build_runner_core/lib/src/generate/build_impl.dart
@@ -35,8 +35,8 @@ import '../util/build_dirs.dart';
 import '../util/constants.dart';
 import 'build_definition.dart';
 import 'build_result.dart';
-import 'create_merged_dir.dart';
 import 'exceptions.dart';
+import 'finalized_assets_view.dart';
 import 'heartbeat.dart';
 import 'options.dart';
 import 'performance_tracker.dart';
@@ -60,7 +60,6 @@ class BuildImpl {
   final Resolvers _resolvers;
   final ResourceManager _resourceManager;
   final RunnerAssetWriter _writer;
-  final Map<String, String> _outputMap;
   final bool _trackPerformance;
   final bool _verbose;
   final BuildEnvironment _environment;
@@ -80,7 +79,6 @@ class BuildImpl {
         _writer = buildDefinition.writer,
         _assetGraph = buildDefinition.assetGraph,
         _resourceManager = buildDefinition.resourceManager,
-        _outputMap = options.outputMap,
         _verbose = options.verbose,
         _environment = buildDefinition.environment,
         _trackPerformance = options.trackPerformance,
@@ -128,7 +126,6 @@ class _SingleBuild {
   final List<BuildPhase> _buildPhases;
   final BuildEnvironment _environment;
   final _lazyPhases = <String, Future<Iterable<AssetId>>>{};
-  final Map<String, String> _outputMap;
   final PackageGraph _packageGraph;
   final BuildPerformanceTracker _performanceTracker;
   final AssetReader _reader;
@@ -152,7 +149,6 @@ class _SingleBuild {
       : _assetGraph = buildImpl._assetGraph,
         _buildPhases = buildImpl._buildPhases,
         _environment = buildImpl._environment,
-        _outputMap = buildImpl._outputMap,
         _packageGraph = buildImpl._packageGraph,
         _performanceTracker = buildImpl._trackPerformance
             ? new BuildPerformanceTracker()
@@ -198,12 +194,8 @@ class _SingleBuild {
       }
     }
     await _resourceManager.disposeAll();
-    if (_outputMap != null && result.status == BuildStatus.success) {
-      if (!await createMergedOutputDirectories(_outputMap, _assetGraph,
-          _packageGraph, _reader, _environment, optionalOutputTracker)) {
-        result = _convertToFailure(result, failureType: FailureType.cantCreate);
-      }
-    }
+    result = await _environment.finalizeBuild(result,
+        new FinalizedAssetsView(_assetGraph, optionalOutputTracker), _reader);
     if (result.status == BuildStatus.success) {
       _logger.info('Succeeded after ${humanReadable(watch.elapsed)} with '
           '${result.outputs.length} outputs '
@@ -213,15 +205,6 @@ class _SingleBuild {
     }
     return result;
   }
-
-  BuildResult _convertToFailure(BuildResult previous,
-          {FailureType failureType}) =>
-      new BuildResult(
-        BuildStatus.failure,
-        previous.outputs,
-        performance: previous.performance,
-        failureType: failureType,
-      );
 
   Future<Null> _updateAssetGraph(Map<AssetId, ChangeType> updates) async {
     await logTimedAsync(_logger, 'Updating asset graph', () async {