+++ build_runner_core/test/environment/simple_prompt.dart
@@ -8,7 +8,8 @@ import 'package:logging/logging.dart';
 import 'package:build_runner_core/src/environment/io_environment.dart';
 
 main() async {
-  var env = new IOEnvironment(new PackageGraph.forThisPackage(), true);
+  var env =
+      new IOEnvironment(new PackageGraph.forThisPackage(), assumeTty: true);
   var result = await env.prompt('Select an option!', ['a', 'b', 'c']);
   Logger.root.onRecord.listen(env.onLog);
   new Logger('Simple Logger').info(result);